+++ novae/model.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Literal
 
 import lightning as L
 import numpy as np
@@ -295,7 +296,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         try:
             model = cls.load_from_checkpoint(artifact_path, map_location=map_location, strict=False, **kwargs)
-        except:
+        except TypeError:
             ckpt_version = torch.load(artifact_path, map_location=map_location).get(Keys.NOVAE_VERSION, "unknown")
             raise ValueError(f"The model was trained with `novae=={ckpt_version}`, but your version is {__version__}")
 
@@ -309,6 +310,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
         adata: AnnData | list[AnnData] | None = None,
         *,
         zero_shot: bool = False,
+        reference: str | int | Literal["all", "largest"] = "largest",
         accelerator: str = "cpu",
         num_workers: int | None = None,
     ) -> None:
@@ -319,6 +321,8 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         Args:
             adata: An `AnnData` object, or a list of `AnnData` objects. Optional if the model was initialized with `adata`.
+            zero_shot: If `True`, the model will be used in zero-shot mode, i.e. without training. In this case, the model will assign each cell to a leaf based on the latent representations.
+            reference: Use only if `zero_shot=True`. Reference slide to use for the new prototypes. Can be the AnnData index, a unique slide id, or one of `["all", "largest"]`.
             accelerator: Accelerator to use. For instance, `'cuda'`, `'cpu'`, or `'auto'`. See Pytorch Lightning for more details.
             num_workers: Number of workers for the dataloader.
         """
@@ -338,12 +342,21 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
                 self._compute_representations_datamodule(adata, datamodule)
 
         if self.mode.zero_shot:
-            latent = np.concatenate([adata.obsm[Keys.REPR][utils.valid_indices(adata)] for adata in adatas])
-            self.swav_head.set_kmeans_prototypes(latent)
-            self.swav_head.reset_clustering(only_zero_shot=True)
+            self.assign_to_kmeans_prototypes(adatas, reference)
 
-            for adata in adatas:
-                self._compute_leaves(adata, None, None)
+    def assign_to_kmeans_prototypes(
+        self, adatas: AnnData | list[AnnData], reference: str | int | Literal["all", "largest"]
+    ):
+        """Compute prototypes based on the latent representations, and assign each cell to a leaf."""
+        adatas = [adatas] if isinstance(adatas, AnnData) else adatas
+        adatas_reference = _get_reference(adatas, reference)
+
+        latent = np.concatenate([adata.obsm[Keys.REPR][utils.valid_indices(adata)] for adata in adatas_reference])
+        self.swav_head.set_kmeans_prototypes(latent)
+        self.swav_head.reset_clustering(only_zero_shot=True)
+
+        for adata in adatas:
+            self._compute_leaves(adata, None, None)
 
     @torch.no_grad()
     def _compute_representations_datamodule(
@@ -417,12 +430,14 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
         weights, thresholds = self.swav_head.queue_weights()
         weights, thresholds = weights.numpy(force=True), thresholds.numpy(force=True)
 
-        where_enough_prototypes = (weights >= thresholds).sum(1) >= self.swav_head.min_prototypes
-        for i in np.where(where_enough_prototypes)[0]:
-            weights[i, weights[i] < thresholds] = 0
-        for i in np.where(~where_enough_prototypes)[0]:
-            indices_0 = np.argsort(weights[i])[: -self.swav_head.min_prototypes]
-            weights[i, indices_0] = 0
+        for i in range(len(weights)):
+            below_threshold_ilocs = np.where(weights[i] < thresholds)[0]
+            if len(thresholds) - len(below_threshold_ilocs) >= self.swav_head.min_prototypes:
+                weights[i, below_threshold_ilocs] = 0
+            else:
+                n_missing = len(thresholds) - self.swav_head.min_prototypes
+                ilocs = below_threshold_ilocs[np.argpartition(weights[i, below_threshold_ilocs], n_missing)[:n_missing]]
+                weights[i, ilocs] = 0
 
         plot._weights_clustermap(weights, self.adatas, list(self.swav_head.slide_label_encoder.keys()), **kwargs)
 
@@ -496,6 +511,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
         self,
         adata: AnnData | list[AnnData],
         *,
+        reference: str | int | Literal["all", "largest"] = "largest",
         accelerator: str = "cpu",
         num_workers: int | None = None,
         min_prototypes_ratio: float = 0.6,
@@ -507,6 +523,7 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         Args:
             adata: An `AnnData` object, or a list of `AnnData` objects. Optional if the model was initialized with `adata`.
+            reference: Reference slide to use for the new prototypes. Can be the AnnData index, a unique slide id, or one of `["all", "largest"]`.
             accelerator: Accelerator to use. For instance, `'cuda'`, `'cpu'`, or `'auto'`. See Pytorch Lightning for more details.
             num_workers: Number of workers for the dataloader.
             min_prototypes_ratio: Minimum ratio of prototypes to be used for each slide. Use a low value to get highly slide-specific or condition-specific prototypes.
@@ -519,14 +536,16 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         assert adata is not None, "Please provide an AnnData object to fine-tune the model."
 
-        datamodule = self._init_datamodule(self._prepare_adatas(adata), sample_cells=Nums.DEFAULT_SAMPLE_CELLS)
-        self.init_slide_queue(adata, min_prototypes_ratio)
-
+        datamodule = self._init_datamodule(
+            self._prepare_adatas(_get_reference(adata, reference)), sample_cells=Nums.DEFAULT_SAMPLE_CELLS
+        )
         latent = self._compute_representations_datamodule(None, datamodule, return_representations=True)
         self.swav_head.set_kmeans_prototypes(latent.numpy(force=True))
 
         self.swav_head._prototypes = self.swav_head._kmeans_prototypes
         del self.swav_head._kmeans_prototypes
+
+        self.init_slide_queue(adata, min_prototypes_ratio)
 
         self.fit(
             adata=adata,
@@ -624,3 +643,34 @@ def _train(
         **trainer_kwargs,
     )
     trainer.fit(model, datamodule=datamodule)
+
+
+def _get_reference(
+    adata: AnnData | list[AnnData], reference: str | int | Literal["all", "largest"]
+) -> AnnData | list[AnnData]:
+    if reference == "all":
+        return adata
+
+    if isinstance(reference, int):
+        assert isinstance(adata, list), "When providing an index, you must provide a list of AnnData objects."
+        return adata[reference]
+
+    if reference == "largest":
+
+        def _select_largest_slide(adata: AnnData):
+            counts = adata.obs[Keys.SLIDE_ID].value_counts()
+            return counts.max(), adata[adata.obs[Keys.SLIDE_ID] == counts.idxmax()]
+
+        if isinstance(adata, AnnData):
+            return _select_largest_slide(adata)[1]
+        else:
+            return max([_select_largest_slide(_adata) for _adata in adata])[1]
+
+    assert isinstance(reference, str), f"Invalid type for `reference`: {type(reference)}"
+
+    adatas = [adata] if isinstance(adata, AnnData) else adata
+    for adata in adatas:
+        if reference in adata.obs[Keys.SLIDE_ID].cat.categories:
+            return adata[adata.obs[Keys.SLIDE_ID] == reference]
+
+    raise ValueError(f"Did not found slide id `{reference}` in the provided AnnData object(s).")